requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.12
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.5
django-bootstrap-form==3.4
django-bootstrap-icons==0.7.9
django-bootstrap4==3.0.1
django-crispy-forms==1.13.0
django-extensions==3.1.3
django-import-export==2.8.0
django-stdimage==5.3.0
django-widget-tweaks==1.4.8
et-xmlfile==1.1.0
filelock==3.0.12
gunicorn==20.1.0
idna==3.3
MarkupPy==1.14
names==0.3.0
odfpy==1.4.1
openpyxl==3.0.10
pbr==5.6.0
Pillow==8.3.1
platformdirs==2.2.0
psycopg2==2.9.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
PyYAML==6.0
reportlab==3.6.3
requests==2.28.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
static3==0.7.0
stevedore==3.3.0
tablib==3.2.1
urllib3==1.26.9
virtualenv==20.7.2
virtualenv-clone==0.5.6
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
xlrd==2.0.1
xlwt==1.3.0
